Add implmentation of gradient descent

+++ ml_hypothesis_gradient.hpp
@@ -9,12 +9,12 @@ class ML_Hypothesis_Gradient : public ML_Hypothesis {
     public:
     ML_Hypothesis_Gradient( std::string filename );
     ~ML_Hypothesis_Gradient( void );
-    void Init( int option1, int option2, int option3, int option4 );
+    void Init( double option1, double option2, double option3, double option4 );
     void Solve( void );
 
     private:
-    int theta0;
-    int theta1;
-    int alpha;
+    double theta0;
+    double theta1;
+    double alpha;
 
 };
